+++ .storybook/preview.js
@@ -1,3 +1,6 @@
+import Theme from '../src/lib/Theme/Theme.svelte';
+import "$lib/style/main.css";
+
 /** @type { import('@storybook/svelte').Preview } */
 const preview = {
 	parameters: {
@@ -7,8 +10,14 @@ const preview = {
 				color: /(background|color)$/i,
 				date: /Date$/i
 			}
-		}
-	}
+		},
+    options: {
+      storySort: {
+        order: ['Introduction', 'Theming/Theme', 'Theming', 'Components', 'Layout', 'Utils', 'Actions', 'Stores'],
+      }
+    }
+	},
+  decorators: [() => Theme],
 };
 
 export default preview;