+++ src/lib/ProjectCredits/ProjectCredits.svelte
@@ -2,7 +2,6 @@
   import Heading from "$lib/HeadingAlt/HeadingAlt.svelte";
   import TextBlock from "$lib/TextBlock/TextBlock.svelte";
   import Block from "$lib/Block/Block.svelte";
-  import "../style/app.css";
 
   /**
    * Heading to display above credits block
@@ -29,7 +28,7 @@
 <!-- Content to render between the heading and the items -->
 <slot name="intro" />
 <Block>
-  <ul>
+  <ul class="credits-list">
     {#each items as item}
       <li>
         <span class="label">{item.label}</span>
@@ -43,7 +42,7 @@
 {/if}
 
 <style>
-  ul {
+  ul.credits-list {
     list-style-type: none;
     padding: 0;
     font-size: var(--font-size-large);